initial:
this project uses p5.js. you can use any program library you need

make a rollercoaster modeling program which can create vertecies to maniuplate the track as segments of curves. the curves can be made with templates including, but not limited to, straight lines, 90 degree circular arcs, 45 degree circular arcs, and an arc of an elliptical shape. the curves can use quadratic bezier curves with a control vertex or any curve model which allows for customization and efficient runtime.

the program should render the coaster as curves in a 3D environment where the user can move around and snap to viewpoints

a HUD panel with textboxes lists the coordinates of each vertex in the rollercoaster. the coordinates can be manually changed. the vertex as a whole, along with any connected control vertecies, can be deleted with a button next to the coordinates